pyproject.toml
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  |  \.tox
  | \.pytest_cache
  | __pycache__
)/
'''

[tool.poetry]
name = "ms-graph-exporter"
version = "0.1.0-rc.1"
description = """A distributed Celery application to export time-domain data 
periodically from Microsoft Graph API into a buffer key-value store."""

license = "MIT"

authors = ["Oleksiy Kuzmenko <oleksiy.kuzmenko@undp.org>"]

repository = "https://github.com/undp/MsGraphExporter"
homepage = "https://github.com/undp/MsGraphExporter"

readme = "README.md"

include = [
  "CHANGELOG.md",
  "CONTRIBUTING.md",
  "LICENSE.md",
  "docker/*",
  "docs/conf.py",
  "docs/*.rst",
  "docs/*.md",
  "tests/*.py",
  "tox.ini",
]

keywords = [
  "development",
  "monitoring",
  "microsoft graph",
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "Natural Language :: English",
  "Operating System :: POSIX :: Linux",
  "Environment :: No Input/Output (Daemon)",
  "Topic :: System :: Monitoring",
]

[tool.poetry.dependencies]
adal = "^1.2"
celery= "^4.3"
celery-redbeat = "^0.13.0"
gevent = "^1.4.0"
python = "^3.6"
redis = "^3.2"
requests = "^2.20.1"
typing = "^3.7"
pyyaml = "^3.13"

# test
pytest = {version = "^5.0.1", optional = true}
pytest-benchmark = {version = "^3.2", extras = ["aspect"], optional = true}
pytest-cov = {version = "^2.7", optional = true}
pytest-dockerc = {version = "^1.0", optional = true}
pytest-instafail = {version = "^0.4.1", optional = true}
pytest-lazy-fixture = {version = "^0.5.2", optional = true}
pytest-random-order = {version = "^1.0", optional = true}
pytest-variables = {version = "^1.7", extras = ["yaml"], optional = true}

# docs
recommonmark = {version = "^0.5.0", optional = true}
sphinx = {version = "^2.0", optional = true}
sphinx_rtd_theme = {version = "^0.4.3", optional = true}
sphinx-autodoc-typehints = {version = "^1.7", optional = true}

# code-lint
flake8 = {version = "^3.7", optional = true}
flake8-import-order = {version = "^0.18.1", optional = true}
flake8-bandit = {version = "^2.1", optional = true}
flake8-blind-except = {version = "^0.1.1", optional = true}
flake8-bugbear = {version = "^19.3", optional = true}
flake8-builtins = {version = "^1.4", optional = true}
flake8-docstrings = {version = "^1.3", optional = true}
flake8-logging-format = {version = "^0.6.0", optional = true}
flake8-mypy = {version = "^17.8", optional = true}
pep8-naming = {version = "^0.8.2", optional = true}
pygments = {version = "^2.4", optional = true}

# code-format
black = {version = "^19.3b0", optional = true}
blacken-docs = {version = "^0.5.0", optional = true}


[tool.poetry.extras]
test = [
  "pytest",
  "pytest-benchmark",
  "pytest-cov",
  "pytest-dockerc",
  "pytest-instafail",
  "pytest-lazy-fixture",
  "pytest-random-order",
  "pytest-variables",
]

docs = [
  "recommonmark",
  "sphinx",
  "sphinx_rtd_theme",
  "sphinx-autodoc-typehints",
]

code-lint = [
  "flake8",
  "flake8-import-order",
  "flake8-bandit",
  "flake8-blind-except",
  "flake8-bugbear",
  "flake8-builtins",
  "flake8-docstrings",
  "flake8-logging-format",
  "flake8-mypy",
  "pep8-naming",
  "pygments",
  "pydocstyle",
]

code-format = [
  "black",
  "blacken-docs",
]

[tool.poetry.dev-dependencies]
pre-commit = "^1.16"
tox = "^3.9"
bump2version = "^0.5.10"
gitchangelog = "^3.0"
pystache = "^0.5.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
